+++ fastrace/benches/compare.rs
@@ -18,7 +18,7 @@ fn init_fastrace() {
     struct DummyReporter;
 
     impl fastrace::collector::Reporter for DummyReporter {
-        fn report(&mut self, _spans: &[fastrace::prelude::SpanRecord]) {}
+        fn report(&mut self, _spans: Vec<fastrace::prelude::SpanRecord>) {}
     }
 
     fastrace::set_reporter(DummyReporter, fastrace::collector::Config::default());